job-info: add job id to update responses
Problem: The responses from update-lookup and update-watch do
not include the job id of the job.  Thus the responses do not
look like the responses from the basic job-info.lookup.  This can
be inconvenient for callers making calls to both.

In all update-lookup and update-watch responses, add the job id
to the response.