resource: add no-update-watch config parameter for testing
Problem: There is no way to test the resource module's fallback to
job-info.lookup from job-info.update-watch when fetching R.

Add a new config parameter, resource.no-update-watch, which causes the
resource module to swap the job-info.update-watch topic string to an
invalid valu when doing the initial fetch of the job's R. This will
then trigger an ENOSYS response from the module, thus simluating an
instance that does not support job-info.update-watch and activating
the fallback code.

Add a debug level log message when falling back to job-info.lookup
so that it can be positively verified that the fallback code was
triggered during testing.