libeventlog: add human readable timestamp to formatter
Problem: The eventlog formatter does not support a timestamp format
that is easily human readable while also making it easy to determine
the date and time of events.

Add a "reltime" or "human" timestamp format, which borrows the same
format as flux-dmesg(1) human readable format (which itself is taken
from the kernel dmesg(8) human readable format).

The reltime format presents a short datetime string for the initial
event and any event that follows from the last "timebreak" by greater
than or equal to one minute. Otherwise, events are given a relative
offset from the last timebreak, e.g.

  [Nov27 18:40] submit userid=6885 urgency=16 flags=0 version=1
  [  +0.011694] validate
  [  +0.022243] depend
  [  +0.022290] priority priority=16

This allows a user to quickly determine when the events occurred and
how far apart they are occurring at a glance.