libeventlog: add color support to eventlog formatter
Problem: Colorized output could really help readability of eventlogs,
but the eventlog formatter class does not currently support color.

Add color support to the eventlog formatter class.

By default, color support is off. To enable, a new function

 eventlog_formatter_colors_init (evf, when);

is called, which takes a de-facto standard string for "when", which
can be one of:

 "auto"   - use color if stdout is a tty
 "never"  - disable color
 "always" - enable color