testsuite: address race in t2204-job-exec-dummy.sh
Problem: t2204-job-exec-dummy.sh was failing reliably on the corona
system recently. It was also noted that it has been failing more often
in CI. The failure was in this simple test:

```
1701268548.313221 submit userid=6885 urgency=16 flags=0 version=1
1701268548.325126 validate
1701268548.335856 depend
1701268548.335901 priority priority=16
1701268548.336890 alloc
1701268548.345470 start
flux-job: eventlog path exec.eventlog not found
not ok 5 - job-exec: job exception kills job shells
```

Git bisect called out the following commit as the proximate cause:

 685f1d0fd testsuite: use flux-cancel not flux-job cancel

And indeed, replacing `flux cancel $id` with `flux job cancel $id`
"fixes" the failure. This is mysterious, because the two commands
should do the same exact thing, and indeed using `FLUX_HANDLE_TRACE=1`
each command sends the exact same RPC. Therefore, a plausible
hypothesis is the failure is due to some kind of race condition,
and somehow the difference in timing between `flux job cancel` and
`flux cancel` causes the test failure.

Since this test uses the "dummy" job shell, which is written in
bash, one theory is that bash is sometimes blocking the SIGTERM
sent by the job-exec module, perhaps before the "tasks" are
launched. Therefore, add an additional wait for the `shell.start`
event in the `guest.exec.eventlog` to ensure the tasks have been
started. With this change the job is canceled successfully but the
finish status of the job is not `15` because the tasks are terminated
not the shell, and the shell exits with exit code `128+15=143`.
Therefore, adjust the test to also accept `143<<8 = 36608` as the
exit code of the terminated job.