testsuite: skip a cmddriver test when testing installed flux(1)
Problem: One of the tests in t1102-cmddriver.t assumes that the `flux`
executable will not be found first in /bin or /usr/bin, but this may
not be the case when FLUX_TEST_INSTALLED_PATH is being used.

Skip the test if flux is found first in /bin or /usr/bin.