testsuite: improve reliability of flux-top queue tests
Problem: The queue output tests in t2801-top-cmd.t often fail in
CI.

The issue seems to be related to the checks for 0 matching lines
at the end of some of the tests. As a guess, sometimes one or more
lines of output match due to timing issues, and so the test fails.

Just remove these overly punctilious checks.