testsuite: avoid inherently racy test in t0015-job-output.py
Problem: A test in t0015-job-output.py attempts to ensure that a
FileNotFoundError is returned from a job output watch when the job
guest namespace has been created, but the output eventlog does
not yet exist. But this test has a race if the job is released
before the output watch RPC is recieved by the job-info module,
and there is no good way to avoid the race during the test since the
`output_event_watch` call is synchronous and called from a thread in
the test

Just remove the racy test so that occasional failures in CI are
avoided.