flux-exec: fix initialization of rank_count
Problem: The global rank_count is unnecessarily initialized in two
different conditionals.

Move intitialization of rank_count to after the target rank set 'ns' is
initialized. This also fixes the rank_count when -x, --exclude is used.

Exit with an error if rank_count is zero instead of silently doing
nothing.