kvs: limit the content of guest commits
Problem: in a multi-user Flux instance, guest users may commit
artibrary content to their private job namespace.

While a job is running, KVS operations from the job owner are
redirected to a private KVS namespace that the job owner may commit
objects to.  This capability supports the job shell running as the guest,
and also allows workflows to use the KVS for synchronization in place of
ad-hoc methods.

When the job transitions to inactive state, the job-exec module deletes
the private namespace, but first grafts its content into the primary
namespace.  This gives the guest user a little more power than may be
expedient.

Limit guest commit/fence operations to: put value, mkdir, and unlink.

Add the commit check directly to the kvs.commit and kvs.fence message
handlers rather than in the lower layers to improve the code's clarity
for auditing purposes.

Drop a test from t1005-kvs-security.t that fails with this change.

Co-authored-by: Jim Garlick <garlick.jim@gmail.com>