flux-exec: simplify --ranks/--exclude option processing
Problem: The determination of target ranks in flux-exec is slightly
more complicated than necessary.

Simplify the iniialization of target ranks in flux-exec by initializing
an idset with all possible ranks, then treat `-r, --ranks` as a set
of ranks to include and `-x, --exclude` as a set to exclude. Emit
an appropriate error if the incalude set or exclude set has ranks
outside the set of all ranks.

Remove tests from t1005-exec.t that target invalid ranks since that
is now caught during this check.