flux-exec: make --exclude ranks relative with --jobid
Problem: When using the flux-exec --jobid option, a user must know the
ranks on which the job is running in order to use the -x, --exclude
option since it requires absolute ranks.

If the --jobid option is in use, make -x, --exclude relative to
the job ranks, such that `-x 0` excludes the first rank of the job,
`-x 1` excludes the second rank, and so on.