flux-terminus: fix possible shutdown race
Problem: There is a potential race in flux-terminus shutdown,
between when the server leaves the Flux reactor and when the service
is unregistered in the broker. If a request message is sent within
this window, then the server cannot respond and the client will hang.

In the terminus server, unregister the service in the broker before
exiting the Flux reactor to close this race.

Fixes #5595