python: handle missing /proc/PID/environ in LSF uri resolver
Problem: There's a small race in the LSF uri resolver if a flux-broker
process for the calling user exits between when it is listed by ps(1)
and the resolver attempts to open the /proc/PID/environ file. In this
case, instead of continuing to try other brokers `flux-uri` aborts with

 flux-uri: ERROR: No such file or directory: '/proc/PID/environ'

Catch the FileNotFound exception when trying to read /proc/PID/environ
and treat this the same as if the broker did not match the LSF jobid.