testsuite: fix possible race in t0005-exec.t
Problem: The 'signal forwarding works' test occasionally fails in ci.

A potential issue is that the test_signal.sh test is run twice,
once with SIGINT and once with SIGTERM, and the sleepready.out
logfile is not cleaned up between the two runs. If the subsequent
waitfail invocation starts before the background process overwrites
the existing sleepready.out, then the test could prematurely send
SIGTERM before flux-exec is ready.

Remove the output file in the test script to avoid this race.