inception: run sharness tests with -d -v
Problem: The inception sharness tess are not run in verbose and debug
mode, which makes diagnosing errors difficult.

Submit tests in the test-inception.sh with 3 different bulksubmit
commands so that `-d -v` can be added to sharness tests.

Use flux-watch(1) to log the output of the tests in real time instead
of the --watch option to bulksubmit.