testsuite: fix potential races in t2233-job-info-update.t
Problem: Several tests in t2233-job-info-update.t check for an initial
number of job update watchers after submitting a job and waiting
for the 'start' event, in order to properly calculate the expected
watcher count for the test. However, the job shell contributes to the
watch count and isn't guaranteed to have installed the update watch
on R until the 'shell.init' event in the exec eventlog. Therefore,
these tests fail occasionally when the update count occurs before
the shell watch.

Wait for 'exec.shell.init' instead of just the 'start' event in these
tests.