testsuite: disable exit-timeout in --stdin-ranks test
Problem: In the --stdin-ranks test in t2500-job-attach.t, a test runs
cat on 2 ranks, but only directs stdin to rank 0. This causes rank
1 to exit immediately and if the test does not finish before 30s,
then the doom plugin kills the job and the test fails.

Set `-o exit-timeout=none` to disable the doom plugin and give the
job the full 60s timeout.