job-list: fix wait state support
Problem: The job-list.list-id RPC nominally supports an optional
job state to wait on for synchronization purposes, but this doesn't
work in all cases. If the job is not unknown or in state NEW in
get_job_by_id(), then the job information is returned immediately
instead of waiting on the requested state.

Modify get_job_by_id() so it waits if a state is requested, but
always return a result immediately if the job is already inactive.

Fixes #5619