libflux: make flux_response_derive() do matchtag
Problem: flux_response_derive() relies on a side-effect of
flux_msg_set_type() to preserve the request matchtag when
building a response message.  The side effect has no tests and
relies somewhat on the way the message proto struct is arranged.

Copy the matchtag from request to response explicitly.  This is
a low cost operation.