libflux: improve message flag interface
Problem: flux_msg_get_flags() and flux_msg_set_flags() are awkward
to use and are limited to flag values that fit in one byte.

These accessors are tagged with the following comment so I think they
are safe to change:

 /* Get/set flags
  * Users should avoid using flux_msg_set_flags(), and instead use the
  * higher level functions that manipulate message flags.  It is exposed
  * mainly for testing.
  */

A cursory check in other framework repos does not turn up any uses of
these functions.

Replace them with:
  flux_msg_has_flag()
  flux_msg_set_flag()
  flux_msg_clear_flag()

which accept an int flag value and require less code to utilize
than the previous functions.

Update users.