job-info: support JSON_DECODE flag
Problem: It can be inconvenient to lookup a key and always have
to convert it to a json object / python dictionary (such as
a jobspec or R).

Solution: Support a new FLUX_JOB_LOOKUP_JSON_DECODE flag that will treat
some special keys (currently jobspec and R) as special lookups and
will returned the values as decoded JSON objects instead of strings.