resource: fetch local hwloc XML from parent if possible
Problem: The resource module always loads local topology via
hwloc_topology_load(3), an expensive operation.

When there is a parent, attempt to fetch the topology XML from the
parent resource module. This is not only more efficient (much more so
on some systems), but also ensures that the initializing subinstance is
using the same hwloc topology used by the parent when a fake topology
is loaded for testing purposes.

If the request fails, or there is no parent instance detected, then
the old behavior is preserved.

Fixes #5611