job-list: add nodelist hostlist to job_data
Problem: In the near future it'd be convenient to do calculations
on the job nodelist, but it often needs to be in a hostlist
data structure for processing.  We'd like to avoid converting the
nodelist to hostlist struct over and over again.

Add a field into the job_data struct to hold a cached version of the
nodelist in a hostlist struct.