testsuite: avoid artificial corefile in test
Problem: In t3306-system-routercrash.t SIGSEGV is intentionally sent
to a broker to test a broker crash, but this could leave a mysterious
corefile in the test directory if the core file ulimit and system
configuration allow it.

Call `ulimit -c 0` before test_under_flux() to avoid the core file.