broker: move background process group to foreground when stopped
Problem: The only interactive commands that can be run as an intial
program are those the broker can detect as a shell. For example, `flux
alloc -n1 vim` fails with "Killing stopped non-interactive process".
This occurs because the broker places rc2 procesess in a separate
process group unless the command is specifically detected as an
interactive shell. When processes in this background process group
attempt to access the terminal they are stopped by SIGTTIM/SIGTTOU,
after which the broker terminates them with "Killing stopped
non-interactive process".

Instead of killing stopped background processes, attempt to bring
them into the foreground if the broker's process group is also in
the foreground. After rc2 exits, restore the broker as the foreground
process group.