flux-job: do not ignore SIGTTIN, SIGTTOU in attach
Problem: flux-job ignores SIGTTIN/SIGTTOU, which prevents it from
being stopped when accessing the terminal in the background. However,
this prevents the broker from detecting that the process wants to
access the terminal when running as an initial program, and instead
of being brought to the foreground, flux-job gets EIO errors when
reading from the terminal and then hangs.

Drop the ignore of SIGTTIN/SIGTTOU in flux-job attach.