broker: runat: always run subprocesses in new process group
Problem: The broker avoids running commands that are detected as an
interactive shell in a new process group to ensure that the command
has access to the terminal, but this is no longer necesssary since
the broker now attempts to bring the background process group into
the foreground if they are stopped.

Always run subprocesses in a new process group in broker/runat.c.