testsuite: add simple expect/send support to runpty
Problem: The runpty.py test script only supports sending input based
on elapsed time, but this is prone to race conditions and may require
long timeouts.

Add a new `--expect` option to `runpty.py` that takes a JSON array of
expect/send entries which can be used to drive input to the command.