testsuite: setup fake hostnames for job list tests
Problem: In the near future we would like to filter jobs based
on nodes in a job's nodelist.  This would be an issue in the current
test setup because test brokers all run under the same host and
it is unknown which nodes jobs actually ran on.

Solution: Setup fake hostnames for the test brokers in
t2260-job-list.t and when necessary, run test jobs on specific
hosts.