kvs: store correct namespace after getroot lookup
Problem: When a getroot is done internally for a KVS lookup, we
wish to store it in the local kvsroot hash.  The namespace
name used for a hash lookup after a getroot request is NOT the namespace
name from the getroot request.  Instead, we incorrectly used the
namespace name from the original KVS lookup request.
This may differ from the intended namespace name.

The result of this may lead to a "infinite loop" where a KVS lookup
needs the root of particular namespace, but the getroot of that
namespace is never retrieved and stored.

Solution: Use the namespace name from a getroot response instaed of
the original KVS lookup request message.

Fixes #5657