testsuite: add issue reproducer for #5657
Problem: In issue #5657, a lookup that contains a symlink to
another namespace would not work due to an internal error with
how namespaces are cached.  The result was a hang on a KVS command
and internally an "infinite loop" in the KVS module.

Add a regression test for this error.