python: Add convert_values member to ConstraintParser class
Problem: Values are always returned as strings by the ConstraintParser
parser, but there are times when another type may be more
suitable. Additionally, there is currently no way to reduce a set
of values to a single element if this is required or beneficial in
the result.

Add a convert_values mapping to the ConstraintParser class. If an
operator is in this dictionary, then the values of a term are passed
to the provided callable, which should return a new list of values
as a result.