flux-filemap: add get --overwrite
Problem: it's too easy for a user to set up a copy operation using
flux filemap where the source and destination files are the same,
because the default is to allow files to be silently overwritten.

Change the default behavior of 'flux filemap get' to fail if a
destination file exists.  Allow this to be overridden with the
--overwrite option (mimicking tar --overwrite).

Helps with #5655