libfilemap: improve libarchive overwrite error message
Problem:  When ARCHIVE_EXTRACT_NO_OVERWRITE is set, the overwrite
error from libarchive-3.6 is "Attempt to write to an empty file".
This is going to be confusing when the file is not empty, such as the
common situation where source and destination of a copy operation are the
same file.

Rewrite that message.