flux-filemap: fix incorrect error message
Problem: flux filemap map reports EBADF instead of EPERM when
open(2) fails but lstat(2) succeeds.

The problem is when fileref_create_ex() calls open(2) with O_NOFOLLOW
and it fails due to EPERM, this is not treated as a fatal error.
Instead, lstat(2) is called with the assumption that it, too, will fail
on a regular file when in fact, it will succeed as long as the containing
directory can be searched.  The code then falls through to mmap(2) but
the file descriptor is not open.

Solution: call lstat(2) only if open(2) with O_NOFOLLOW fails with ELOOP.