broker: allow single-user rexec to rank 0
Problem: remote rexec to rank 0 is not permitted, but it would
be useful to allow shell plugins to run things on rank 0 in order
to utilize leader-only services like 'flux filemap map'.

Remote rexec is restricted in order to prevent the flux user from
remotely executing arbitrary programs on a restricted management node.
This is not a concern for other Flux instances such as batch jobs.

Change the rexec authorization callback to allow remote access on rank 0
if the 'access.allow-guest-user' TOML key is unset or false.

Update tests.

Fixes #5676