build: drop jsonschema requirement
Problem: Installation of flux-core currently requires the Python
jsonschema package, but the jsonschema validator is slow, a maintenance
burden, and has no users outside of the testsuite.

Remove the jsonschema job validator, schemas, and all users in the
flux-core codebase.

The jsonschema packages are still provided in the test Dockerfiles
since dependent flux-framework projects (such as flux-sched) may
still depend on this package.