broker: unset FLUX_PROXY_REMOTE environment variable
Problem: flux-proxy(1) sets the FLUX_PROXY_REMOTE environment variable
as a hint to flux_attr_get(3) that the parent-uri attribute should be
translated to a remote uri so that commands running under the proxy
can use that URI when they are actually on a remote system. However,
this causes incorrect behavior within a broker, where the attempt to
translate to a remote URI is not only incorrect, but could result in
an RPC and deadlock in attr_cache_to_json() while loading a module.

Unset FLUX_PROXY_REMOTE in the broker since the start of new broker
most definitely means subsequent processes are no longer under the
influence a remote flux-proxy(1).

Fixes #5679