libflux: fix cache miss for parent-uri when FLUX_PROXY_REMOTE is set
Problem: When FLUX_PROXY_REMOTE is set, `flux_attr_get(3)` will
substitute parent-remote-uri for a request for the parent-uri
attribute.  However, if the function fails to find parent-remote-uri
in the attr cache, it always issues an RPC to fetch the parent-uri
attribute from which parent-remote-uri is derived, instead of checking
the cache for that attr.

Fix the code to check for orig_name (parent-uri) in the cache before
resorting to an RPC.