job-list: add match context infrastructure
Problem: In the near future, some initialization will have to be
done for constraint matching.  However, no information can currently
be passed to the list_constraint_create() function.

Add infrastructure to allow a match context to be initialized on
job-list load and can be passed to list_constraint_create().  While
not presently used, it will allow easier adjustments to constraint
matching in the future.