librlist: fix incorrect size passed to hwloc_topology_set_xmlbuffer
Problem: When an hwloc topology is initialized from XML in rhwloc.c,
the size of the entire XML buffer include the terminating NUL character
is passed to hwloc_topology_topology_set_xmlbuffer(). This causes hwloc
to return an error with errno set to EINVAL with hwloc version 2.9.3.

Passing the XML string length instead of the buffer size does not
appear to cause any issues on previous hwloc versions, so just do that.