job-list: allow job_match() to return error
Problem: The job_match() function currently returns a bool, indicating
if a job matches a constraint or does not.  In the future, we will need
to be able to return an error from job_match(), but that is not currently
possible.

Refactor job_match() to return an int instead of a bool.  It now returns
1 for a match, 0 for not, and -1 on error.  In addition, have job_match()
take a flux_error_t pointer as an argument, in preparation for future
error message returns.