libsubprocess: simplify fbuf callback
Problem: callback semantics for fbuf are too complicated.

The only users are the ev_fbuf watchers which need a notification
when the buffer transitions from full or from empty, and do not require
a size threshold.

Consolidate callbacks to one "notify" callback.  The callback can check
the buffer for the desired condition.

Update users and tests.