testsuite: fix queue stats check after reload in t2260-job-list
Problem: The `flux job stats` per-queue tests after a reload of the
job-list module will not have any data on queues for which no jobs
have been submitted. However, a test in t2260-job-list.t assumes an
entry for such a queue will exist with all values 0.

Remove the tests for the queue with no jobs and instead check that
the entry is empty.