testsuite: fix uses of jq with no input
Problem: Older versions of jq(1) (pre 1.7) exit with success when
the input is empty, even if a conditional check is provided, i.e.

   echo | jq '.test.foo == 42'

succeeds instead of fails. This has led to several instances of
undetected missing input to jq in the testsuite.

Fix these cases to provide the expected test JSON input.