librlist: workaround xml buffer size issue in some hwloc versions
Problem: With some versions of libxml, the XML load function does not
tolerate a buffer length that includes the trailing NUL character.
In hwloc 2.10.0, hwloc_topology_set_xmlbuffer() was changed to
decrement the length before passing to the underlying XML call, so
this effectively means that hwloc now *requires* the NUL character
be included in length, but current librlist code does not include it
to workaround the problem in v2.10.

Since there is no way to determine if NUL should or should not be
included in hwloc_topology_set_xmlbuffer(), include the NUL by default,
but retry without it if the first call fails.

Fixes #5689