libsubprocess: make refcount private
Problem: flux_subprocess_ref() and flux_subprocess_unref() are
not used but are in the public API.

Rename
flux_subprocess_ref() => subprocess_incref()
flux_subprocess_unref() => subprocess_decref()

and change flux_subprocess_destroy() from a preprocessor macro to
an actual function.

Update unit tests.

Update one user that was forced to use _unref() because it required the
destructor to be assinged to a function pointer.