job-list: limit constraint comparisons
Problem: Job-list constraints are not limited in their size.  A nefarious
user could cause a DoS of the job-list service by sending an extremely
large constraint request.

Solution: To limit constraints, introduce the concept of constraint
comparisons.  All comparison "checks", such as seeing if a job was submitted
by a specific user, are counted.  When the maximum number of comparisons is
reached, an error is returned to the user.  The user can try again by
limiting the number of jobs they are searching through, such as only
searching active vs inactive jobs or if jobs were submitted since a certain time.

For the time being, hard code the maximum number of comparisons to 100
thousand.

Fixes #5669