libterminus: encode data as base64 when json_pack() fails
Problem: Encoding pty data with jansson json_pack() or equivalent
occasionally fails. This is most likely because of the documented
caveat:

 All Unicode codepoints U+0000 through U+10FFFF are allowed in
 string values. However, U+0000 is allowed in object keys only for
 length-aware functions.

Attempt to encode pty data with base64 when packing the data as utf-8
fails. In this case, add an optional (to avoid breaking the protocol)
"encoding" key set to "base64".

Since there is now shared code between the client and server, export
some helper functions for encoding/decoding pty "data" messages to
pty.c with support for the fallback to base64.