testsuite: make `flux alloc vi` test more reliable
Problem: The `flux alloc vi works` test in t2712-python-cli-alloc.t
occasionally fails because the vi statusline displaying the filename
never appears.

Actually, we just want to ensure that vi started and the file was
loaded, so change the expected output to be the file contents, which
should always be displayed unless there is a real issue.