python: fix handling of base64 output in JobWatcher
Problem: If there is binary output, encoded in base64 in the job
output eventlog, the JobWatcher class and thus flux-watch(1) throws
errors about bytes not being encodable as JSON.

First, ensure that base64 encoded output is decoded to utf-8 with
surrogates, so that the data can be encoded back to binary without
data loss.

Then, in the JobWatcher class, ensure all output streams are capable
of handling the surrogates (and utf-8) by reconfiguring them (in
python 3.7+) or reopening them with the proper error handling.

Fixes #5702