python: add optional constraint parameter to JobList
Problem: There is no way to pass extra constraints via the JobList
interface.

Add an optional constraint parameter to the JobList initializer. This
paramter takes a string in constraint query syntax which is parsed
by an instance of JobListConstraintParser. The result is then joined
with ``and`` to any other constraints before being passed to the
job-list module.