job-ingest: handle fluid_init() failure
Problem: if fluid_init() fails, job-ingest could issue
duplicate/invalid job IDs.

When fluid_init() fails, job-ingest logs an error but continues
on.  Since the "fluid_generator" is just a 64-bit integer, the
fluids generated after that might be the same as those issued
on rank 0, and may start from an incorrect clock offset.

fluid_init() is expected to fail on rank > 16383.
It could also fail in the unlikely event that clock_gettime() fails.

When fluid_init() fails, abort the module.