job-ingest: don't fail on rank > 16383
Problem: fluid_init() fails on rank > 16383 due to a design
limitation on the number of concurrent fluid generators.

If fluid_init() fails on one of these high ranks, log a message at
debug level and return 0, so the instance can continue with ingest
requests on that rank being handled upstream.

Fixes #5708