job-list: update queue_stats_lookup()
Problem: The internal queue_stats_lookup() function looks up
a stats structure for a queue and creates one if it is missing.
But in a few cases, this makes no sense.  For example, when a
job is being purged, it should be (generally) impossible that
queue stats don't exist.

Add a boolean indicating if the queue stats should be created
or not in queue_stats_lookup().  In a few cases where it's undesireable for
queue_stats_lookup() to return no queue stats, output some
debugging error info.

Also, have queue_stats_lookup() take a queue name instead of a
"struct job" as input, as the queue name is what is actually necessary.