job-list: initialize queue stats
Problem: Job queue stats are generated in job-list when a job
is submitted into it.  If a job queue never has a job submitted
to it, no job stats exist for the queue.  This isn't what users
would expect, they should expect queue stats to be 0 for the queue.

Solution: Read the flux config and initialize queue stats to 0 for
any queues that have been configured.

Fixes #5688