testsuite: update job list queue stat tests
Problem: A test in t2260-job-list.t assumes there will be no
job queue stats if a job has never been submitted to it.  That
behavior was recently changed.

Update the test to check for non-empty queue stats for queues
that should have no jobs "submitted" to it.  Also add a new
"nosubmitqueue" to the test for additional coverage.