job-list/test: update for comparisons checks
Problem: The match_context struct is not initialized with
the new max_comparisons field for unit tests.  In addition,
now that the values in the ctx will actually be used, it should
not sit on the stack, otherwise we might read corrupted data.

Create a global match ctx for all list constraint creations.  Set
the max_comparisons to 0 for unlimited in the unit tests.