shell: add flux PMI dir to LD_LIBRARY_PATH
Problem: Flux libpmi.so and libpmi2.so are installed to a directory
outside the default ld.so search path, so Slurm's tend to be found
first by Flux jobs.

Add the PMI library install directory to LD_LIBRARY_PATH in the task
environment when the Flux simple PMI plugin is active.

Fixes #5714