test: add simple MPI scaling script to src/test/scaling
Problem: A reusable script to drive MPI scaling tests has been written
but is not available so it can be conveniently reused.

Add src/test/scaling/mpi-scale.sh - a script that uses the timing
mode of t/mpi/hello to generate a table of MPI init/barrier/finalize
timing results as the number of tasks scale up.