cmd: add flux-hostlist
Problem: It would be convenient to have a utility for working with
Flux RFC 29 Hostlists -- to create a hostfile for a job or instance,
or for comparing the hosts assigned to failing jobs -- but no such
utility currently exists.

Add flux-hostlist(1), a utility that takes one or more hostlist
"sources" on the command line and supports several options to query
or modify the result.

Supported sources include:

 instance: use instance hostlist attribute

 local (the default): use the hosts assigned to FLUX_JOB_ID if set,
  o/w the hostlist attribute of the enclosing instance.

 avail[able]: use the hosts currently available for job scheduling.
  i.e. exclude drained and down hosts.

 stdin or '-': read hosts from stdin.

 jobid: hosts assigned to job

 or a literal host or list of hosts.

The form of the result may be manipulated by one of the set operation
options (union, intersect, minus, or xor), and the tool may be used to
print the result (limited or modified by options), count its members,
sort, etc.