python: support raw description formatter with help_formatter()
Problem: There is currently no way to get raw description and
epilog formatting with the FluxHelpFormatter class returned from
help_formatter().

Add a FluxRawDescriptionHelpFormatter class that inherits from both
FluxHelpFormatter and RawDescriptionHelpFormatter for this purpose.
Return this class instead of FluxHelpFormatter if a new optional
raw_description parameter is set in the call to help_formatter().