flux-shell: abort if taskmap=block is used with a value
Problem: If --taskmap=block:foo is used, the job shell silently
ignores the value 'foo'. This could lead to a job running with
the wrong taskmap if the user intended some other taskmap scheme.
It would be better to abort and notify the user something's wrong.

Check if a value is specified with `--taskmap=block` and abort the
shell if so.