shell: output: make kvs output limit configurable
Problem: The job shell truncates output at 10M to workaround current
KVS limitations with very large eventlogs in possibly long-running
system instances. However, this is unnecessarily limiting in
single-user subinstances where it is most convenient to have all job
output folded into the output of a batch job.

Convert the KVS output limit to a configurable value set via the job
shell option `output.limit`. Keep the default limit of 10M if the
userid of the enclosing instance broker does not match the current
userid, and disable the limit completely if the broker and shell
userids match.

Update shell output truncation tests that would now fail with the
new behavior. Ensure output limit is not enforced in the single user
test instance.

Note that this allows a user to disable the limit in a multiuser instance
as well via `-o output.limit=0`, but this was possible anyway since
the limit is not really enforced by the instance (a user could build
a custom job shell for instance).