flux-archive: new command
Problem: flux-filemap only works on rank 0.

Add a new command that stores RFC 37 archives in the KVS instead
of in the content module, and by default copies large file content
into the KVS content store so the source files can be on any rank.

If the --mmap option is specified when the archive is created,
the content of large files is mmapped as it was with flux-filemap(1).
This option only works on rank 0 and does not support --preserve.
It is potentially useful when large files such as VM images need to
be propagated from rank 0 to other ranks, and copying them into the
KVS content store is impractical.  The archive metadata still goes
to the KVS.

The archive metadata (which includes small file data) is stored to
the KVS under one key.  To minimize the size of the archive metadata,
the small file threshold is reduced from 4K in flux-filemap(1) to 1K,
and remains user-adjustable so performance can be tuned.

The archive metadata includes blobrefs that have no RFC 11 KVS metadata
linkage, which means they may be garbage collected by a dump/restore.
To avoid this, a --preserve option specified when the archive is created
causes an archive.<name>_blobs directory to be created, containing a key
for each blobref in the archive.

If a name is unspecified, "main" is used, which translates to a KVS key
of "archive.main" and will match the default name in the shell stage-in
plugin once it is updated.  Archives are stored in the primary KVS
namespace by default, unless --no-force-primary is specified, and then
FLUX_KVS_NAMESPACE is respected.