libtaskmap: support decode of raw taskmaps
Problem: libtaskmap and `flux job taskmap` can encode to raw taskmaps,
but cannot decode tham. Besides being asymmetric, it is sometimes
much easier to encode a manual taskmap in raw rather than RFC 34 form,
so this should be supported.

Add support for decoding raw taskmaps in taskmap_decode(3). The
code uses a heuristic to determine if the taskmap is in raw form
by only decoding if no characters in the set "({[]})" appear in the
input string.