shell/stage-in: use new KVS archive format
Problem: the shell stage-in plugin only works with mmapped
files.

Rename the stage-in.tags option to stage-in.keys and change
the code so it reads the main archive from the KVS.

Rework the stage-in sharness test so it does most of the same
things as before but using the new tooling.