flux-perilog-run: do not drain ranks when job is canceled
Problem: When a user cancels a job while the prolog is running,
the prolog is canceled with a SIGTERM. This causes the flux-exec
tasks spawned by the flux-perilog-run --exec-per-rank option to be
terminated, which causes those ranks to be drained. Thus, a user can
cause nodes to be drained by canceling their job.

If an exec-per-rank process is terminated by a signal but wasn't
canceled by flux-perilog-run itself (i.e. proc.canceled is not True),
then assume the prolog was terminated and do not drain ranks. That is,
only drain ranks when subprocess returncode is between 0 and 129 to
account for shell-reported signal (128+n) or actual signal (-n).

Fixes #5734