modules/content: drop content.mmap-list RPC
Problem: flux-archive stores file lists in the KVS, so the
mechansim for listing mmapped files via the content module is
no longer needed.

Drop the RPC handler and associated functions.
Drop the code for making filerefs available via the content cache
as that was only required for 'blobref' mode of this RPC.
Drop filemap_mmap_list() from libfilemap.