python: allow {id.f58plain} to be used in output formats
Problem: The f58plain encoding of jobids could be useful in some
situations with flux-jobs and other tools that support {id.<encoding>}
output format, but this encoding is currently not available since
it does not appear in `job_fields_to_attrs()` or the JobInfoFormat
headings dict.

Add `id.f58plain` where necessary to make it available as an option
in flux-jobs etc.